Disable small caps ligatures for the moment
The lookup table of substitution of SmallCapitals made the conversion
script generate warnings saying that not all entries had the same type.

So, I'm commenting this for the moment.

I suspect that this is due to the fact that the SmallCapitals part may
be a single-substitution only table, while we are trying to substitute
two symbols (namely, f.sc and i.sc with fi).

I don't have the manual handy here to check, though.