Initial addition of Old Style Figures from Gaël
Imported the OSF from Gaël with the hint of Khaled. Thank you for both
of you.

CAUTION: no substitution table has been added to reference the glyphs
yet. This is one of the next steps.

Unfortunately, it seems that there are no OSF in upright/regular, so it
is currently not available.  Perhaps with a little bit of help from
Fontforge some of the other OSF versions could be made for the regular
shape?

What I have in mind is, perhaps, a small dose of these:

* unslant the OSF from italic and italic bold
* un-embold the OFS from the bold ones
* interpolate them, if looking promising
* make fixes by hand

Any other transformations?