Clean up the old style figures from the upright bold font
There is still a lot of work to be done.

Both the 6 and the 9 look very bad and the 8 would be much more charming
if it had its bottom end bigger than its upper half (of course, to an
extent that doesn't deviate too much from the general "feel" of the
family).