Fix the automated generation of slanted fonts
Just discovered what made the slant script generate internal errors in
fontforge: the fact that I was selecting the glyphs.

Once that step is removed, there are no messages of problems being
output to the user and the script succeeds in creating the fonts.

Also make sure that we are importing the feature file and keep an eye on
the italic angle: -16, copied from the regular italic font.