Try to make generation of slanted fonts automatic
I'm not sure how to name the fonts so that they can be picked up by
fontspec, but this should not be a problem as long as we are not (yet)
making a release version of the fonts.

The only reference that I get handy is the thread at
   http://www.tug.org/pipermail/xetex/2008-April/009535.html
but that was quite some time ago.